backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary>=2.9.9
celery==5.3.6
django-celery-results==2.5.1
redis==5.0.1
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
